+++ example/lib/presentation/src/widgets/src/scroll_view/src/ac_scroll_view_data_source.dart
@@ -39,6 +39,7 @@ abstract class ACScrollViewDataSource<T> extends ChangeNotifier {
   /// Загружает следующий элемент (от текущего) для анимации навигации.
   /// Возвращает элемент или null если навигация невозможна.
   T? loadAfter();
+
 }
 
 /// Реализация [ACScrollViewDataSource] по умолчанию.